requirements.txt
garminconnect==0.2.28
numpy==1.22.4
pandas==2.2.1
python-dotenv==1.0.1
python_lichess==0.10
pytz==2023.3.post1
Requests==2.31.0
scikit_learn==1.4.0
xgboost==1.5.0
matplotlib==3.5.1
boto3
garth
modal
supabase==2.3.4